+++ src/riscv64/bindings.rs
@@ -630,6 +630,10 @@ pub type __wsum = __u32;
 pub type __poll_t = ::std::os::raw::c_uint;
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct user_regs_struct {
     pub pc: ::std::os::raw::c_ulong,
     pub ra: ::std::os::raw::c_ulong,
@@ -1592,6 +1596,10 @@ fn bindgen_test_layout_kvm_sregs() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_config {
     pub isa: ::std::os::raw::c_ulong,
     pub zicbom_block_size: ::std::os::raw::c_ulong,
@@ -1688,6 +1696,10 @@ fn bindgen_test_layout_kvm_riscv_config() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_core {
     pub regs: user_regs_struct,
     pub mode: ::std::os::raw::c_ulong,
@@ -1729,6 +1741,10 @@ fn bindgen_test_layout_kvm_riscv_core() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_csr {
     pub sstatus: ::std::os::raw::c_ulong,
     pub sie: ::std::os::raw::c_ulong,
@@ -1869,6 +1885,10 @@ fn bindgen_test_layout_kvm_riscv_csr() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_aia_csr {
     pub siselect: ::std::os::raw::c_ulong,
     pub iprio1: ::std::os::raw::c_ulong,
@@ -1965,6 +1985,10 @@ fn bindgen_test_layout_kvm_riscv_aia_csr() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_smstateen_csr {
     pub sstateen0: ::std::os::raw::c_ulong,
 }
@@ -1996,6 +2020,10 @@ fn bindgen_test_layout_kvm_riscv_smstateen_csr() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_timer {
     pub frequency: __u64,
     pub time: __u64,
@@ -2128,6 +2156,10 @@ pub const KVM_RISCV_SBI_EXT_ID_KVM_RISCV_SBI_EXT_MAX: KVM_RISCV_SBI_EXT_ID = 11;
 pub type KVM_RISCV_SBI_EXT_ID = ::std::os::raw::c_uint;
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_riscv_sbi_sta {
     pub shmem_lo: ::std::os::raw::c_ulong,
     pub shmem_hi: ::std::os::raw::c_ulong,
@@ -6387,6 +6419,10 @@ fn bindgen_test_layout_kvm_vapic_addr() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_mp_state {
     pub mp_state: __u32,
 }
@@ -7549,6 +7585,10 @@ fn bindgen_test_layout_kvm_reg_list() {
 }
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, PartialEq)]
+#[cfg_attr(
+    feature = "serde",
+    derive(zerocopy::AsBytes, zerocopy::FromBytes, zerocopy::FromZeroes)
+)]
 pub struct kvm_one_reg {
     pub id: __u64,
     pub addr: __u64,